+++ src/main/kotlin/com/taali/api/dto/teacher/TeacherDashboardStats.kt
@@ -1,7 +1,7 @@
 package com.taali.api.dto.teacher
 
 data class TeacherDashboardStats(
-    val totalStudents: Int,
+    val totalStudents: Long,
     val totalClasses: Int,
     val upcomingClasses: Int,
     val attendanceRate: Double
